# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ALERYAF HUB API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: getInventory
      tags: [inventory]
      summary: Get current inventory state
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Inventory list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
  /inventory/import:
    post:
      operationId: importInventory
      tags: [inventory]
      summary: Import daily inventory from Excel/CSV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryImportResult"
...
components:
  schemas:
...
    InventoryItem:
      type: object
      properties:
        itemId:
          type: integer
        itemCode:
          type: string
        itemName:
          type: string
        itemNameAr:
          type: string
        category:
          type: string
        importDate:
          type: string
        importedQuantityKg:
          type: number
        unitDisplay:
          type: string
        costPerKg:
          type: number
        totalValue:
          type: number
        currentBalanceKg:
          type: number
        currentValue:
          type: number
        currency:
          type: string
          enum: [TRY, USD]
        openingBalance:
          type: number
        soldQuantity:
          type: number
        currentStock:
          type: number
        minStock:
          type: number
        unitCostTry:
          type: number
        unitCostUsd:
          type: number
        inventoryValueTry:
          type: number
        inventoryValueUsd:
          type: number
        isLowStock:
          type: boolean
      required:
        - itemId
        - itemCode
        - itemName
        - openingBalance
        - soldQuantity
        - currentStock
        - isLowStock
    InventoryImportInput:
      type: object
      properties:
        importDate:
          type: string
        warehouseId:
          type: integer
        rows:
          type: array
          items:
            $ref: "#/components/schemas/InventoryImportRow"
      required:
        - importDate
        - warehouseId
        - rows
    InventoryImportRow:
      type: object
      properties:
        itemCode:
          type: string
        rawName:
          type: string
        quantity:
          type: number
        unit:
          type: string
        costTry:
          type: number
        costUsd:
          type: number
        totalValue:
          type: number
      required:
        - quantity
    InventoryImportResult:
      type: object
      properties:
        importId:
          type: integer
        importDate:
          type: string
        rowsProcessed:
          type: integer
        rowsMatched:
          type: integer
        rowsUnmatched:
          type: integer
        unmatchedItems:
          type: array
          items:
            type: string
      required:
        - importId
        - importDate
        - rowsProcessed
        - rowsMatched
        - rowsUnmatched
        - unmatchedItems
...