# lib/api-zod/src/generated/types/inventoryImportRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ALERYAF HUB API
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryImportRow {
  itemCode?: string;
  rawName?: string;
  quantity: number;
  unit?: string;
  costTry?: number;
  costUsd?: number;
  totalValue?: number;
}